+++ flexmeasures/cli/tests/test_data_add.py
@@ -1,4 +1,7 @@
 import pytest
+import json
+import os
+
 
 from flexmeasures.cli.tests.utils import to_flags
 from flexmeasures.data.models.annotations import (
@@ -6,7 +9,10 @@ from flexmeasures.data.models.annotations import (
     AccountAnnotationRelationship,
 )
 from flexmeasures.data.models.data_sources import DataSource
+from flexmeasures.data.models.time_series import Sensor
+
 from flexmeasures.cli.tests.utils import get_click_commands
+from flexmeasures.utils.time_utils import server_now
 
 
 @pytest.mark.skip_github
@@ -87,3 +93,121 @@ def test_cli_help(app):
         result = runner.invoke(cmd, ["--help"])
         assert result.exit_code == 0
         assert "Usage" in result.output
+
+
+@pytest.mark.skip_github
+def test_add_reporter(app, db, setup_dummy_data, reporter_config_raw):
+    """
+    The reporter aggregates input data from two sensors (both have 200 data points)
+    to a two-hour resolution.
+
+    The command is run twice:
+        - The first run is for ten hours, so you expect five results.
+            - start and end are defined in the configuration: 2023-04-10T00:00 -> 2023-04-10T10:00
+            - this step uses 10 hours of data -> outputs 5 periods of 2 hours
+        - The second is run without timing params, so the rest of the data
+            - start is the time of the latest report value
+            - end is the time of latest input data value
+            - this step uses 190 hours of data -> outputs 95 periods of 2 hours
+    """
+
+    from flexmeasures.cli.data_add import add_report
+
+    sensor1, sensor2, report_sensor = setup_dummy_data
+    report_sensor_id = report_sensor.id
+
+    # Running the command with start and end values.
+
+    runner = app.test_cli_runner()
+
+    cli_input_params = {
+        "sensor-id": report_sensor_id,
+        "reporter-config": "reporter_config.json",
+        "reporter": "PandasReporter",
+        "start": "2023-04-10T00:00:00 00:00",
+        "end": "2023-04-10T10:00:00 00:00",
+        "output-file": "test.csv",
+    }
+
+    cli_input = to_flags(cli_input_params)
+
+    # run test in an isolated file system
+    with runner.isolated_filesystem():
+
+        # save reporter_config to a json file
+        with open("reporter_config.json", "w") as f:
+            json.dump(reporter_config_raw, f)
+
+        # call command
+        result = runner.invoke(add_report, cli_input)
+
+        print(result)
+
+        assert result.exit_code == 0  # run command without errors
+
+        assert "Reporter PandasReporter found" in result.output
+        assert "Report computation done." in result.output
+
+        # Check report is saved to the database
+
+        report_sensor = Sensor.query.get(
+            report_sensor_id
+        )  # get fresh report sensor instance
+
+        stored_report = report_sensor.search_beliefs(
+            event_starts_after=cli_input_params.get("start").replace(" ", "+"),
+            event_ends_before=cli_input_params.get("end").replace(" ", "+"),
+        )
+
+        assert (
+            stored_report.values.T == [1, 2 + 3, 4 + 5, 6 + 7, 8 + 9]
+        ).all()  # check values
+
+        assert os.path.exists("test.csv")  # check that the file has been created
+        assert (
+            os.path.getsize("test.csv") > 100
+        )  # bytes. Check that the file is not empty
+
+    # Running the command without without timing params (start-offset/end-offset nor start/end).
+    # This makes the command default the start time to the date of the last
+    # value of the reporter sensor and the end time as the current time.
+
+    previous_command_end = cli_input_params.get("end").replace(" ", "+")
+
+    cli_input_params = {
+        "sensor-id": report_sensor_id,
+        "reporter-config": "reporter_config.json",
+        "reporter": "PandasReporter",
+        "output-file": "test.csv",
+        "timezone": "UTC",
+    }
+
+    cli_input = to_flags(cli_input_params)
+
+    with runner.isolated_filesystem():
+
+        # save reporter_config to a json file
+        with open("reporter_config.json", "w") as f:
+            json.dump(reporter_config_raw, f)
+
+        # call command
+        result = runner.invoke(add_report, cli_input)
+
+        print(result)
+
+        assert result.exit_code == 0  # run command without errors
+
+        assert "Reporter PandasReporter found" in result.output
+        assert "Report computation done." in result.output
+
+        # Check if the report is saved to the database
+        report_sensor = Sensor.query.get(
+            report_sensor_id
+        )  # get fresh report sensor instance
+
+        stored_report = report_sensor.search_beliefs(
+            event_starts_after=previous_command_end,
+            event_ends_before=server_now(),
+        )
+
+        assert len(stored_report) == 95